-N/--skip-newer accepts a float now.

+++ generate_tiles.py
@@ -736,11 +736,11 @@ def parse_args():
     parser.add_argument('-X', '--skip-existing', dest='skip_existing', default=False,
                         action='store_true')
     # TODO: newer than input_file
-    parser.add_argument('-N', '--skip-newer',    dest='skip_newer', default=None,
-                        type=int, metavar='DAYS')
-    parser.add_argument(      '--missing-as-new',  dest='missing_as_new', default=False, action='store_true',
-                        help="missing tiles in a meta tile count as newer, so we don't re-render metatiles with empty tiles.")
-    parser.add_argument('-E', '--empty',         dest='empty',     default='skip',
+    parser.add_argument('-N', '--skip-newer',      dest='skip_newer', default=None,
+                        type=float, metavar='DAYS')
+    parser.add_argument(      '--missing-as-new',  dest='missing_as_new', default=False,
+                        action='store_true', help="missing tiles in a meta tile count as newer, so we don't re-render metatiles with empty tiles.")
+    parser.add_argument('-E', '--empty',           dest='empty',     default='skip',
                         choices=('skip', 'link', 'write'))
 
     parser.add_argument(      '--debug',         dest='debug',     default=False, action='store_true')
